fix - 매개변수 타입을 Long 으로 변경

+++ data/src/main/java/com/yapp/growth/data/api/GrowthApi.kt
@@ -19,7 +19,7 @@ interface GrowthApi {
     suspend fun getAllFixedPlanList(): List<FixedPlanResponseImpl>
 
     @GET("/api/promises/{promiseId}")
-    suspend fun getFixedPlan(@Path("promiseId") pId: Int): FixedPlanResponseImpl
+    suspend fun getFixedPlan(@Path("promiseId") pId: Long): FixedPlanResponseImpl
 
     @GET("/api/promisings/session/{uuid}")
     suspend fun getCreateTimeTable(@Path("uuid") uuid: String): CreateTimeTableResponseImpl
